Added methods and filtered games
